\name{Version1-package}
\alias{Version1-package}
\alias{Version1}
\docType{package}
\title{
\packageTitle{Version1}
}
\description{
\packageDescription{Version1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Version1}
\packageIndices{Version1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Version1}

Maintainer: \packageMaintainer{Version1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
